backend/django/pyproject.toml
[tool.poetry]
name = "django-api-authentication"
version = "0.1.0"
description = ""
authors = ["Uadson Feitosa <uadsonpy@gmail>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<4.0"
django = "^5.1.4"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
django-cors-headers = "^4.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
